% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_class}
\alias{check_class}
\title{check_class}
\usage{
check_class(obj, cls)
}
\arguments{
\item{obj}{Any R object.}

\item{cls}{Character string. Describes a class type, e.g. "matrix", "list",
"vector", etc...}
}
\description{
check_class
}
