% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{discrete_plot_trans}
\alias{discrete_plot_trans}
\title{Plot the synchronous test results.}
\usage{
discrete_plot_trans(tr, dir, name, fdr, annot, num_perm, trans_hit_vals,
  trans_perm_obs_results, tr_and_pheno_hi_conf, geno_confidence,
  g_trans_edges, p_trans_edges, snp_in_gene, prefix, grouped_logical)
}
\arguments{
\item{tr}{Phylo.}

\item{dir}{Directory where to save plots.}

\item{name}{Prefix in plot file name.}

\item{fdr}{Numeric. False discovery rate. Between 0 and 1.}

\item{annot}{Deprecated.}

\item{num_perm}{Numeric. Number of permutations.}

\item{trans_hit_vals}{Dataframe. Nrows = number of genotypes. Ncol = 1.
Corrected p-values for each genotype tested.}

\item{trans_perm_obs_results}{List of many results.  $hit_pvals. Character.
P-val for each genotype. Length = number of tested genotypes.
$permuted_count. List of vectors. 1 vector for each tested genotype. Length
of each vector = number of permuations. $observed_overlap. Integer vector.
Length = number of tested genotypes.}

\item{tr_and_pheno_hi_conf}{Vector of logicals. TRUE = high confidence. FALSE
= low confidence. Length = Nedge(tree).}

\item{geno_confidence}{List of vectors. Length of individual vector =
Nedge(tree). Genotype high confidence edges. Either 1 (high confidence) or
0 (low confidence).}

\item{g_trans_edges}{List of vectors. Length of individual vector =
Nedge(tree). Genotype transition edges. Either 1 (transition) or 0 (no
transition).}

\item{p_trans_edges}{Vector. Length = Nedge(tree). Transitions marked as 1,
not transition marked as 0.}

\item{snp_in_gene}{Either NULL or Table of integers where each entry
corresponds to one genotype.}
}
\value{
Plots printed into one pdf.
}
\description{
Plot the synchronous test results.
}
