% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth}
\alias{growth}
\title{Continuous phenotype}
\format{A matrix with 7 rows and 1 columns. Each row name corresponds to a
  tree tip.
\describe{
  \item{growth}{Dummy growth rates}
 }}
\usage{
growth
}
\description{
An example continous phenotype matrix.
}
\keyword{datasets}
