% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hogwash.R
\name{hogwash}
\alias{hogwash}
\title{Main hogwash function}
\usage{
hogwash(pheno, geno, tree, file_name = paste(Sys.Date(), "_hogwash", sep
  = ""), dir = ".", perm = 10000, fdr = 0.15, bootstrap = 0.7,
  group_genotype_key = NULL)
}
\arguments{
\item{group_genotype_key}{}
}
\description{
Main hogwash function
}
