% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_ann_colors}
\alias{make_ann_colors}
\title{Create object to annotate columns in the significant hit results.}
\usage{
make_ann_colors(geno_matrix)
}
\arguments{
\item{geno_matrix}{}
}
\value{

}
\description{
Create object to annotate columns in the significant hit results.
}
