% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_manhattan_plot}
\alias{make_manhattan_plot}
\title{Create a manhattan plot of GWAS hit p-values.}
\usage{
make_manhattan_plot(outdir, geno_pheno_name, pval_hits, alpha,
  trans_or_recon)
}
\arguments{
\item{trans_or_recon}{}
}
\description{
Create a manhattan plot of GWAS hit p-values.
}
