% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_continuous_phenotype}
\alias{plot_continuous_phenotype}
\title{Plot the continuous phenotype on tree with the phenotype coloring the branches.}
\usage{
plot_continuous_phenotype(tr, pheno_vector, pheno_anc_rec)
}
\arguments{
\item{pheno_anc_rec}{}
}
\value{

}
\description{
Plot the continuous phenotype on tree with the phenotype coloring the branches.
}
