% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_significant_hits}
\alias{plot_significant_hits}
\title{Plot continuous phenotype results.}
\usage{
plot_significant_hits(disc_cont, tr, fdr, dir, name, pval_all_transition,
  pheno_vector, annot, perm, results_all_trans, pheno_anc_rec,
  geno_reconstruction, geno_confidence, geno_transition, geno,
  pheno_recon_ordered_by_edges, tr_and_pheno_hi_conf, all_trans_sig_hits)
}
\arguments{
\item{all_trans_sig_hits}{}
}
\value{

}
\description{
Plot continuous phenotype results.
}
