% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_tree_with_colored_edges}
\alias{plot_tree_with_colored_edges}
\title{Plot a phylogenetic tree with certain edges highlighted.}
\usage{
plot_tree_with_colored_edges(tr, edges_to_highlight, geno_confidence,
  edge_color_na, edge_color_bright, title, annot, trans_or_recon, index)
}
\arguments{
\item{index}{}
}
\value{

}
\description{
Plot a phylogenetic tree with certain edges highlighted.
}
