% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{report_phylogenetic_signal}
\alias{report_phylogenetic_signal}
\title{Report best fitting model for the phenotype and tree}
\usage{
report_phylogenetic_signal(pheno, tree)
}
\arguments{
\item{pheno}{Matrix. One column matrix. Row.names correspond to tree$tip.labels.}

\item{tree}{Phylo.}
}
\description{
Report best fitting model for the phenotype and tree
}
