% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_binary_original.R
\name{run_binary_original}
\alias{run_binary_original}
\title{Run discrete convergence test based on PhyC algorithm.}
\usage{
run_binary_original(args)
}
\arguments{
\item{args}{}
}
\value{

}
\description{
Run discrete convergence test based on PhyC algorithm.
}
