% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp_gene_key}
\alias{snp_gene_key}
\title{Key}
\format{A matrix with 8 rows and 2 columns.
\describe{
  \item{SNP}{SNP identifiers}
  \item{GENE}{Gene names}
 }}
\usage{
snp_gene_key
}
\description{
An example mapping SNPs to genes.
}
\keyword{datasets}
