% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp_genotype}
\alias{snp_genotype}
\title{Binary genotype}
\format{
A matrix with 7 rows and 8 columns. Rows correspond to isolates and
  tree tips. Columns correspond to genotypes.
\describe{
  \item{SNP1}{SNP1 present encoded as 1, absent encoded as 0}
 }
}
\usage{
snp_genotype
}
\description{
An example binary genotype matrix.
}
\keyword{datasets}
