% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{Phylogenetic tree}
\format{A rooted phylogenetic tree with 7 tips and 6 internal nodes.
\describe{
  \item{tip.labels}{Names of each tree tip}
  \item{node.lables}{Bootstrap values for each node}
 }}
\usage{
tree
}
\description{
An example tree with 7 tips.
}
\keyword{datasets}
