% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{discrete_plot_trans}
\alias{discrete_plot_trans}
\title{Plot the synchronous test results.}
\usage{
discrete_plot_trans(tr, dir, name, fdr, annot, num_perm, trans_hit_vals,
  trans_perm_obs_results, tr_and_pheno_hi_conf, geno_confidence,
  g_trans_edges, p_trans_edges, snp_in_gene, prefix, grouped_logical)
}
\arguments{
\item{snp_in_gene}{}
}
\value{

}
\description{
Plot the synchronous test results.
}
